[ENH] Refactor Benchmarking Classes
